Split delaunay and voronoid nodes implementation.
